p5 (Bipartite Graph)/graph_tests.py
import unittest
from graph import *

class TestList(unittest.TestCase):

    def test_01(self):
        g = Graph('test1.txt')
        self.assertEqual(g.conn_components(), [['v1', 'v2', 'v3', 'v4', 'v5'], ['v6', 'v7', 'v8', 'v9']])
        self.assertTrue(g.is_bipartite())
        self.assertEqual(g.get_vertices(), ['v1', 'v2', 'v3', 'v4', 'v5', 'v6', 'v7', 'v8', 'v9'])
        
    def test_02(self):
        g = Graph('test2.txt')
        self.assertEqual(g.conn_components(), [['v1', 'v2', 'v3'], ['v4', 'v6', 'v7', 'v8']])
        self.assertFalse(g.is_bipartite())

    def test_03(self):
        g = Graph("test1.txt")
        self.assertEqual(g.graph["v1"], g.get_vertex("v1"))
        self.assertFalse(g.get_vertex("v20"))
        self.assertEqual(['v1', 'v2', 'v3', 'v4', 'v5', 'v6', 'v7', 'v8', 'v9'], g.get_vertices())

    def test_raise_exception(self):
        with self.assertRaises(Exception):
            g = Graph("yessir<3")


if __name__ == '__main__':
   unittest.main()
